requirements.txt
anyio==3.7.1
asgiref==3.7.2
bcrypt==4.0.1
certifi==2023.7.22
click==8.1.6
dnspython==2.4.0
exceptiongroup==1.1.2
fastapi==0.65.2
h11==0.14.0
httpcore==0.17.3
idna==3.4
passlib==1.7.4
pydantic==1.10.2
PyJWT==2.8.0
pymongo==4.4.1
python-dotenv==1.0.0
python-multipart==0.0.5
six==1.16.0
sniffio==1.3.0
starlette==0.14.2
typing_extensions==4.7.1
uvicorn==0.15.0
